Splitter opp hardkoding av tekster i flere filer og legger på riktige typer
