+++ mocks/testdata/sanity/felles/felles.ts
@@ -1,0 +1,38 @@
+import {
+    IFellesTekstInnhold,
+    IHjelpeteksterForInputTekstInnhold,
+    IKanIkkeBrukeSoeknadTekstinnhold,
+} from '../../../../src/frontend/typer/sanity/tekstInnhold';
+import { lagLocaleRecordBlock } from '../lagSanityObjekter';
+
+import { formateringsfeilmeldingerTekstinnhold } from './formateringsfeilmeldinger';
+import { frittståendeOrdTekstinnhold } from './frittståendeOrd';
+import { modalerTekstinnhold } from './modaler/modaler';
+import { navigasjonTekstinnhold } from './navigasjon';
+
+const vedlikeholdsarbeidTekstinnhold = {
+    vedlikeholdTittel: lagLocaleRecordBlock(),
+    vedlikeholdBroedtekst: lagLocaleRecordBlock(),
+    vedlikeholdVeileder: lagLocaleRecordBlock(),
+};
+
+const kanIkkeBrukeSoeknadTekstinnhold: IKanIkkeBrukeSoeknadTekstinnhold = {
+    brukPDFKontantstoette: lagLocaleRecordBlock(),
+};
+
+const hjelpeteksterForInputTekstInnhold: IHjelpeteksterForInputTekstInnhold = {
+    datoformatHjelpetekst: lagLocaleRecordBlock(),
+    datoformatPlaceholder: lagLocaleRecordBlock(),
+    manedformatPlaceholder: lagLocaleRecordBlock(),
+    velgLandPlaceholder: lagLocaleRecordBlock(),
+};
+
+export const fellesTekstInnhold: IFellesTekstInnhold = {
+    frittståendeOrd: frittståendeOrdTekstinnhold,
+    modaler: modalerTekstinnhold,
+    navigasjon: navigasjonTekstinnhold,
+    formateringsfeilmeldinger: formateringsfeilmeldingerTekstinnhold,
+    vedlikeholdsarbeid: vedlikeholdsarbeidTekstinnhold,
+    kanIkkeBrukeSoeknad: kanIkkeBrukeSoeknadTekstinnhold,
+    hjelpeteksterForInput: hjelpeteksterForInputTekstInnhold,
+};
