+++ mocks/testdata/sanity/felles/modaler/leggTilBarn.ts
@@ -1,0 +1,15 @@
+import { ILeggTilBarnTekstinnhold } from '../../../../../src/frontend/typer/sanity/modaler/leggTilBarn';
+import { lagLocaleRecordBlock, lagSanitySpørsmålDokument } from '../../lagSanityObjekter';
+
+export const leggTilBarnTekstinnhold: ILeggTilBarnTekstinnhold = {
+    tittel: lagLocaleRecordBlock(),
+    leggTilKnapp: lagLocaleRecordBlock(),
+    fjernKnapp: lagLocaleRecordBlock(),
+    erBarnetFoedt: lagSanitySpørsmålDokument(),
+    ikkeFoedtAlert: lagLocaleRecordBlock(),
+    barnetsNavnSubtittel: lagLocaleRecordBlock(),
+    fornavn: lagSanitySpørsmålDokument(),
+    etternavn: lagSanitySpørsmålDokument(),
+    foedselsnummerEllerDNummer: lagSanitySpørsmålDokument(),
+    foedselsnummerAlert: lagLocaleRecordBlock(),
+};
