+++ mocks/testdata/sanity/felles/modaler/startPåNytt.ts
@@ -1,0 +1,9 @@
+import { IStartPåNyttModal } from '../../../../../src/frontend/typer/sanity/modaler/startPåNytt';
+import { lagLocaleRecordBlock } from '../../lagSanityObjekter';
+
+export const startPåNytt: IStartPåNyttModal = {
+    startNySoeknadKnapp: lagLocaleRecordBlock(),
+    startPaaNyttInfo: lagLocaleRecordBlock(),
+    startPaaNyttTittel: lagLocaleRecordBlock(),
+    startPaaNyttAvbryt: lagLocaleRecordBlock(),
+};
