+++ mocks/testdata/sanity/felles/navigasjon.ts
@@ -1,0 +1,15 @@
+import { INavigasjonTekstinnhold } from '../../../../src/frontend/typer/sanity/tekstInnhold';
+import { lagLocaleRecordString } from '../lagSanityObjekter';
+
+export const navigasjonTekstinnhold: INavigasjonTekstinnhold = {
+    startKnapp: lagLocaleRecordString(),
+    fortsettKnapp: lagLocaleRecordString(),
+    startPaaNyttKnapp: lagLocaleRecordString(),
+    duMaaRetteOppFoelgende: lagLocaleRecordString(),
+    avbrytSoeknad: lagLocaleRecordString(),
+    tilbakeKnapp: lagLocaleRecordString(),
+    gaaVidereKnapp: lagLocaleRecordString(),
+    sendSoeknadKnapp: lagLocaleRecordString(),
+    slettSoeknadKnapp: lagLocaleRecordString(),
+    fortsettSenereKnapp: lagLocaleRecordString(),
+};
