+++ mocks/testdata/sanity/sider/dokumentasjon.ts
@@ -1,0 +1,87 @@
+import { IDokumentasjonTekstinnhold } from '../../../../src/frontend/components/SøknadsSteg/Dokumentasjon/innholdTyper';
+import {
+    BeskrivelseSanityApiNavn,
+    TittelSanityApiNavn,
+} from '../../../../src/frontend/typer/dokumentasjon';
+import { lagLocaleRecordString, lagLocaleRecordBlock } from '../lagSanityObjekter';
+
+export const dokumentasjonTekstinnhold: IDokumentasjonTekstinnhold = {
+    dokumentasjonTittel: lagLocaleRecordBlock(),
+    dokumentasjonGuide: lagLocaleRecordBlock(),
+    dokumentasjonGuideVedleggskrav: lagLocaleRecordBlock(),
+    dokumentasjonGuideIngenVedleggskrav: lagLocaleRecordBlock(),
+    // Info innledning
+    forLangTidDokumentasjon: lagLocaleRecordBlock(),
+    vedleggskravTittel: lagLocaleRecordBlock(),
+    vedleggskrav: lagLocaleRecordBlock(),
+    ingenVedleggskravTittel: lagLocaleRecordBlock(),
+    ingenVedleggskrav: lagLocaleRecordBlock(),
+    manglerDokumentasjonSpoersmaalTittel: lagLocaleRecordBlock(),
+    manglerDokumentasjonSpoersmaal: lagLocaleRecordBlock(),
+    // Bilde scanning guide
+    slikTarDuEtGodtBildeExpand: lagLocaleRecordString(),
+    slikTarDuEtGodtBildeTittel: lagLocaleRecordBlock(),
+    slikTarDuEtGodtBilde: lagLocaleRecordBlock(),
+    etterDuHarTattBildetTittel: lagLocaleRecordBlock(),
+    etterDuHarTattBildet: lagLocaleRecordBlock(),
+    vaerTryggNaarDuTarBildeTittel: lagLocaleRecordBlock(),
+    vaerTryggNaarDuTarBilde: lagLocaleRecordBlock(),
+    // Bra og dårlige eksempler
+    braOgDaarligeTittel: lagLocaleRecordBlock(),
+    bra: lagLocaleRecordBlock(),
+    daarlig: lagLocaleRecordBlock(),
+    fyllerHeleBildet: lagLocaleRecordBlock(),
+    ikkeTattOvenfra: lagLocaleRecordBlock(),
+    ikkeRiktigRetning: lagLocaleRecordBlock(),
+    skyggePaaDokumentet: lagLocaleRecordBlock(),
+    // Knapper og checkbox
+    sendtInnTidligere: lagLocaleRecordBlock(),
+    forMange: lagLocaleRecordString(),
+    feilFiltype: lagLocaleRecordString(),
+    forStor: lagLocaleRecordString(),
+    bildetForLite: lagLocaleRecordString(),
+    noeGikkFeil: lagLocaleRecordString(),
+    lastOppKnapp: lagLocaleRecordString(),
+    slippFilenHer: lagLocaleRecordString(),
+    slett: lagLocaleRecordString(),
+    stottedeFiltyper: lagLocaleRecordString(),
+    maksFilstorrelse: lagLocaleRecordString(),
+    maksAntallFiler: lagLocaleRecordString(),
+    lastOppFiler: lagLocaleRecordString(),
+    filtypeFeilmelding: lagLocaleRecordString(),
+    filstorrelseFeilmelding: lagLocaleRecordString(),
+    antallFilerFeilmelding: lagLocaleRecordString(),
+    ukjentFeilmelding: lagLocaleRecordString(),
+    velgFil: lagLocaleRecordString(),
+    velgFiler: lagLocaleRecordString(),
+    draOgSlippFilenHer: lagLocaleRecordString(),
+    draOgSlippFilerHer: lagLocaleRecordString(),
+    filopplastingDeaktivert: lagLocaleRecordString(),
+    filopplastingDeaktivertFilerErUnderOpplastning: lagLocaleRecordString(),
+    filopplastingDeaktivertMaksAntallFiler: lagLocaleRecordString(),
+    lastOppFilenPaNytt: lagLocaleRecordString(),
+    slettFilen: lagLocaleRecordString(),
+    lasterNed: lagLocaleRecordString(),
+    lasterOpp: lagLocaleRecordString(),
+    // Vedlegg - titler
+    [TittelSanityApiNavn.bekreftelsePaaAdopsjonTittel]: lagLocaleRecordBlock(),
+    [TittelSanityApiNavn.annenDokumentasjon]: lagLocaleRecordBlock(),
+    [TittelSanityApiNavn.avtaleOmDeltBostedTittel]: lagLocaleRecordBlock(),
+    [TittelSanityApiNavn.bekreftelseFraBarnevernetTittel]: lagLocaleRecordBlock(),
+    [TittelSanityApiNavn.bekreftelsePaaAtBarnBorSammenMedDegTittel]: lagLocaleRecordBlock(),
+    [TittelSanityApiNavn.meklingsattestTittel]: lagLocaleRecordBlock(),
+    [TittelSanityApiNavn.dokumentasjonPaaSeparasjonSkilsmisseEllerDoedsfallTittel]:
+        lagLocaleRecordBlock(),
+    [TittelSanityApiNavn.vedtakOmOppholdstillatelseTittel]: lagLocaleRecordBlock(),
+    // Vedlegg - beskrivelser
+    [BeskrivelseSanityApiNavn.dokumentasjonPaaSeparasjonSkilsmisseEllerDoedsfall]:
+        lagLocaleRecordBlock(),
+    [BeskrivelseSanityApiNavn.vedtakOmOppholdstillatelse]: lagLocaleRecordBlock(),
+    [BeskrivelseSanityApiNavn.bekreftelsePaaAdopsjonBarnetrygd]: lagLocaleRecordBlock(),
+    [BeskrivelseSanityApiNavn.bekreftelsePaaAtBarnBorSammenMedDeg]: lagLocaleRecordBlock(),
+    [BeskrivelseSanityApiNavn.avtaleOmDeltBosted]: lagLocaleRecordBlock(),
+    [BeskrivelseSanityApiNavn.meklingsattest]: lagLocaleRecordBlock(),
+    [BeskrivelseSanityApiNavn.bekreftelseFraBarnevernetBarnetrygd]: lagLocaleRecordBlock(),
+    [BeskrivelseSanityApiNavn.lastOppSenereISoknad]: lagLocaleRecordBlock(),
+    [BeskrivelseSanityApiNavn.annenDokumentasjonBeskrivelse]: lagLocaleRecordBlock(),
+};
