+++ mocks/testdata/sanity/sider/eøsForBarn.ts
@@ -1,0 +1,53 @@
+import { IEøsForBarnTekstinnhold } from '../../../../src/frontend/components/SøknadsSteg/EøsSteg/Barn/innholdTyper';
+import {
+    lagLocaleRecordBlock,
+    lagLocaleRecordString,
+    lagSanitySpørsmålDokument,
+} from '../lagSanityObjekter';
+
+export const eøsForBarnTekstinnhold: IEøsForBarnTekstinnhold = {
+    eoesForBarnTittel: lagLocaleRecordBlock(),
+    eoesForBarnTittelUtenFlettefelt: lagLocaleRecordBlock(),
+    eosForBarnGuide: lagLocaleRecordBlock(),
+    idNummerBarn: lagSanitySpørsmålDokument(),
+    slektsforhold: lagSanitySpørsmålDokument(),
+    hvilkenRelasjon: lagSanitySpørsmålDokument(),
+    borMedAndreForelder: lagSanitySpørsmålDokument(),
+    borMedOmsorgsperson: lagSanitySpørsmålDokument(),
+    hvorBorBarnet: lagSanitySpørsmålDokument(),
+    subtittelAndreForelder: lagLocaleRecordBlock(),
+    idNummerAndreForelder: lagSanitySpørsmålDokument(),
+    hvorBorAndreForelder: lagSanitySpørsmålDokument(),
+    paagaaendeSoeknadYtelseAndreForelder: lagSanitySpørsmålDokument(),
+    hvilketLandSoektYtelseAndreForelder: lagSanitySpørsmålDokument(),
+    oppgittIkkeBorFastSammenMedDeg: lagLocaleRecordBlock(),
+    hvaHeterOmsorgspersonen: lagSanitySpørsmålDokument(),
+    slektsforholdOmsorgsperson: lagSanitySpørsmålDokument(),
+    hvilkenRelasjonOmsorgsperson: lagSanitySpørsmålDokument(),
+    idNummerOmsorgsperson: lagSanitySpørsmålDokument(),
+    hvorBorOmsorgsperson: lagSanitySpørsmålDokument(),
+    paagaaendeSoeknadYtelseOmsorgsperson: lagSanitySpørsmålDokument(),
+    hvilketLandSoektYtelseOmsorgsperson: lagSanitySpørsmålDokument(),
+    arbeidNorgeAndreForelder: lagSanitySpørsmålDokument(),
+    arbeidNorgeAndreForelderGjenlevende: lagSanitySpørsmålDokument(),
+    pensjonNorgeAndreForelder: lagSanitySpørsmålDokument(),
+    pensjonNorgeAndreForelderGjenlevende: lagSanitySpørsmålDokument(),
+    utbetalingerAndreForelder: lagSanitySpørsmålDokument(),
+    utbetalingerAndreForelderGjenlevende: lagSanitySpørsmålDokument(),
+    ytelseFraAnnetLandAndreForelder: lagSanitySpørsmålDokument(),
+    ytelseFraAnnetLandAndreForelderGjenlevende: lagSanitySpørsmålDokument(),
+    arbeidUtenforNorgeOmsorgsperson: lagSanitySpørsmålDokument(),
+    arbeidNorgeOmsorgsperson: lagSanitySpørsmålDokument(),
+    pensjonUtlandOmsorgsperson: lagSanitySpørsmålDokument(),
+    pensjonNorgeOmsorgsperson: lagSanitySpørsmålDokument(),
+    utbetalingerOmsorgsperson: lagSanitySpørsmålDokument(),
+    ytelseFraAnnetLandOmsorgsperson: lagSanitySpørsmålDokument(),
+
+    /* Slektsforhold valgalternativ */
+    valgalternativSlektsforholdPlaceholder: lagLocaleRecordString(),
+    valgalternativAnnenRelasjon: lagLocaleRecordString(),
+    valgalternativAnnenFamilierelasjon: lagLocaleRecordString(),
+    valgalternativForelder: lagLocaleRecordString(),
+    valgalternativOnkelTante: lagLocaleRecordString(),
+    valgalternativBesteforelder: lagLocaleRecordString(),
+};
