+++ mocks/testdata/sanity/sider/omBarnet.ts
@@ -1,0 +1,44 @@
+import { IOmBarnetTekstinnhold } from '../../../../src/frontend/components/SøknadsSteg/OmBarnet/innholdTyper';
+import {
+    lagLocaleRecordBlock,
+    lagLocaleRecordString,
+    lagSanitySpørsmålDokument,
+} from '../lagSanityObjekter';
+
+export const omBarnetTekstinnhold: IOmBarnetTekstinnhold = {
+    omBarnetTittel: lagLocaleRecordBlock(),
+    omBarnetTittelUtenFlettefelt: lagLocaleRecordBlock(),
+    omBarnetGuide: lagLocaleRecordBlock(),
+    barnetsAndreForelder: lagLocaleRecordString(),
+    opplystFosterbarn: lagLocaleRecordBlock(),
+    opplystInstitusjon: lagLocaleRecordBlock(),
+    institusjonIUtlandetCheckbox: lagLocaleRecordBlock(),
+    institusjonNavn: lagSanitySpørsmålDokument(),
+    institusjonAdresse: lagSanitySpørsmålDokument(),
+    institusjonPostnummer: lagSanitySpørsmålDokument(),
+    institusjonStartdato: lagSanitySpørsmålDokument(),
+    institusjonSluttdato: lagSanitySpørsmålDokument(),
+    institusjonUkjentSluttCheckbox: lagLocaleRecordBlock(),
+    opplystBarnOppholdUtenforNorge: lagLocaleRecordBlock(),
+    planlagtBoSammenhengendeINorge: lagSanitySpørsmålDokument(),
+    opplystFaarHarFaattEllerSoektYtelse: lagLocaleRecordBlock(),
+    paagaaendeSoeknadYtelse: lagSanitySpørsmålDokument(),
+    hvilketLandYtelse: lagSanitySpørsmålDokument(),
+    svaralternativSammeSomAnnenForelder: lagLocaleRecordBlock(),
+    svaralternativAnnenForelder: lagLocaleRecordBlock(),
+    hvemErBarnSinAndreForelder: lagSanitySpørsmålDokument(),
+    navnAndreForelder: lagSanitySpørsmålDokument(),
+    foedselsnummerDnummerAndreForelder: lagSanitySpørsmålDokument(),
+    foedselsdatoAndreForelder: lagSanitySpørsmålDokument(),
+    arbeidUtenforNorgeAndreForelder: lagSanitySpørsmålDokument(),
+    pensjonUtlandAndreForelderGjenlevende: lagSanitySpørsmålDokument(),
+    pensjonUtlandAndreForelder: lagSanitySpørsmålDokument(),
+    boddSammenMedAndreForelder: lagSanitySpørsmålDokument(),
+    naarFlyttetFraAndreForelder: lagSanitySpørsmålDokument(),
+    bosted: lagLocaleRecordString(),
+    bostedInfo: lagLocaleRecordBlock(),
+    borBarnFastSammenMedDeg: lagSanitySpørsmålDokument(),
+    deltBosted: lagSanitySpørsmålDokument(),
+    faarEllerHarFaattYtelseFraAnnetLand: lagSanitySpørsmålDokument(),
+    arbeidUtenforNorgeAndreForelderGjenlevende: lagSanitySpørsmålDokument(),
+};
