+++ src/frontend/utils/testing.tsx
@@ -10,7 +10,7 @@ import { ESvar } from '@navikt/familie-form-elements';
 import { HttpProvider } from '@navikt/familie-http';
 import { type Ressurs, RessursStatus } from '@navikt/familie-typer';
 
-import { hentTekstInnhold } from '../../../mocks/testdata/sanity';
+import { hentTekstInnhold } from '../../../mocks/testdata/sanity/sanity';
 import norskeTekster from '../assets/lang/nb.json' with { type: 'json' };
 import { UtenlandsoppholdSpørsmålId } from '../components/Felleskomponenter/UtenlandsoppholdModal/spørsmål';
 import {
